Added FAT filesystem commands
 For traversing directories and READING from the filesystem 'ls', 'cd', 'pwd' and 'cat. Also added visual code configuration project files for Lab20 console application.

+++ Lab20 Mass Storage/boards/peripherals/dwc/transfer.c
@@ -639,7 +639,7 @@ void TransferStageDataAttach(TransferStageData *transfer, u32 channel,
     assert (transfer->bufferPointer != 0);
   else
     transfer->bufferPointer = &transfer->TempBuffer;
-  assert (((u32) transfer->bufferPointer & 3) == 0);
+  assert (((uintptr_t)transfer->bufferPointer & 3) == 0);
 
   if (transfer->splitTransaction)
   {
